Use dependency on qrcodegen crate
