+++ src/commands/qrcode.rs
@@ -1,6 +1,6 @@
-use crate::lib::qr::print_qr;
 use crate::lib::{read_from_file, AnyhowResult};
 use clap::Parser;
+use qrcodegen::{QrCodeEcc, QrCode};
 
 #[derive(Parser)]
 pub struct QRCodeOpts {
@@ -23,4 +23,27 @@ pub fn exec(opts: QRCodeOpts) -> AnyhowResult {
         print_qr(&string);
     }
     Ok(())
+}
+
+// Prints the given QrCode object to the console.
+pub fn print_qr(text: &str) {
+    let errcorlvl: QrCodeEcc = QrCodeEcc::Medium; // Error correction level
+
+    // Make and print the QR Code symbol
+    let qr: QrCode = QrCode::encode_text(text, errcorlvl).unwrap();
+
+    let border: i32 = 4;
+    for y in -border / 2..qr.size() / 2 + border / 2 + 1 {
+        for x in -border..qr.size() + border {
+            let c = match (!qr.get_module(x, y * 2), !qr.get_module(x, y * 2 + 1)) {
+                (true, true) => '█',
+                (true, false) => '▀',
+                (false, true) => '▄',
+                (false, false) => ' ',
+            };
+            print!("{0}", c);
+        }
+        println!();
+    }
+    println!();
 }
